refactor(#251): modify folder route
