+++ src/admins/banned-user/repositories/banned-user.repository.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { CreateBannedUserBodyDto } from '@src/admins/dtos/create-banned-user-body.dto';
+import { CreateBannedUserBodyDto } from '@src/admins/banned-user/dtos/create-banned-user-body.dto';
 import { BannedUser } from '@src/admins/banned-user/entities/banned-user.entity';
 import { EntityManager, FindOneOptions } from 'typeorm';
