+++ src/admins/banned-user/services/banned-users.service.ts
@@ -6,7 +6,7 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { BannedUserRepository } from '@src/admins/banned-user/repositories/banned-user.repository';
-import { CreateBannedUserBodyDto } from '@src/admins/dtos/create-banned-user-body.dto';
+import { CreateBannedUserBodyDto } from '@src/admins/banned-user/dtos/create-banned-user-body.dto';
 import { UserService } from '@src/users/services/user.service';
 import { DataSource } from 'typeorm';
 import { BannedUserDto } from '@src/admins/banned-user/dtos/banned-user.dto';
@@ -14,7 +14,7 @@ import { UserStatus } from '@src/users/constants/user-status.enum';
 import { ADMIN_ERROR_CODE } from '@src/constants/error/admin/admin-error-code.constant';
 import { AdminException } from '@src/http-exceptions/exceptions/admin-exception';
 import { UserRole } from '@src/users/constants/user-role.enum';
-import { BannedUserPageQueryDto } from '@src/admins/dtos/banned-user-page-query.dto';
+import { BannedUserPageQueryDto } from '@src/admins/banned-user/dtos/banned-user-page-query.dto';
 import { BannedUser } from '@src/admins/banned-user/entities/banned-user.entity';
 import { QueryHelper } from '@src/helpers/query.helper';
 import { plainToInstance } from 'class-transformer';
@@ -148,7 +148,7 @@ export class BannedUsersService {
   async findOneByUserId(userId: number): Promise<BannedUserDto> {
     const existBannedUser = await this.bannedUserRepository.findOne({
       where: { bannedUserId: userId },
-      order: { bannedAt: SortOrder.Desc },
+      order: { bannedAt: SortOrder.DESC },
     });
 
     return existBannedUser ? new BannedUserDto(existBannedUser) : null;
